+++ .test_durations
@@ -79,16 +79,10 @@
   "tests/e2e/test_dock_playground.py::test_snap_two_floating_panels[chromium]": 0.8646563740912825,
   "tests/e2e/test_dock_playground.py::test_tab_reorder[chromium]": 1.4585882080718875,
   "tests/e2e/test_dock_playground.py::test_tab_tear_out[chromium]": 0.8735517079476267,
-  "tests/e2e/test_dock_playground_above_height.py::test_dock_above_single_panel_splits_height_evenly[chromium]": 1.1966556245461106,
-  "tests/e2e/test_dock_playground_above_height.py::test_dock_above_two_columns_spans_and_preserves_widths[chromium]": 1.688972083851695,
   "tests/e2e/test_dock_playground_area.py::test_drag_area_tab_out_and_merge_with_parent[chromium]": 0.9087898333091289,
   "tests/e2e/test_dock_playground_area.py::test_drag_floating_panel_into_area[chromium]": 0.7608674170915037,
   "tests/e2e/test_dock_playground_area.py::test_drag_snapped_stack_into_area_collapses_to_tabs[chromium]": 1.2103720002342016,
   "tests/e2e/test_dock_playground_area.py::test_drag_window_over_own_area_no_self_merge[chromium]": 1.2111523328348994,
-  "tests/e2e/test_dock_playground_cascade.py::test_vertical_cascade_pushes_through[chromium]": 1.119480666005984,
-  "tests/e2e/test_dock_playground_column_handle.py::test_column_handle_floats_whole_stack[chromium]": 1.8819819160271436,
-  "tests/e2e/test_dock_playground_column_handle.py::test_sandwiched_minimized_column_renders_vertical_strip[chromium]": 2.849509332096204,
-  "tests/e2e/test_dock_playground_column_handle.py::test_split_preview_does_not_wipe_region_background[chromium]": 2.0015935401897877,
   "tests/e2e/test_dock_playground_dropzones.py::test_drag_tab_back_uses_live_strip_geometry[chromium]": 1.6892866669222713,
   "tests/e2e/test_dock_playground_dropzones.py::test_grip_bar_splits_above_only_that_panel[chromium]": 1.581924291094765,
   "tests/e2e/test_dock_playground_dropzones.py::test_region_span_preview_is_a_thin_line[chromium]": 1.4777660409454256,
@@ -106,6 +100,13 @@
   "tests/e2e/test_dock_playground_fuzz.py::test_random_drags_never_throw_or_strand_windows[chromium-6]": 4.165040998952463,
   "tests/e2e/test_dock_playground_fuzz.py::test_random_drags_never_throw_or_strand_windows[chromium-7]": 4.280728416983038,
   "tests/e2e/test_dock_playground_fuzz.py::test_random_drags_never_throw_or_strand_windows[chromium-8]": 4.261961957672611,
+  "tests/e2e/test_dock_playground_handles.py::test_column_handle_floats_whole_stack[chromium]": 1.8819819160271436,
+  "tests/e2e/test_dock_playground_handles.py::test_sandwiched_minimized_column_renders_vertical_strip[chromium]": 2.849509332096204,
+  "tests/e2e/test_dock_playground_handles.py::test_split_preview_does_not_wipe_region_background[chromium]": 2.0015935401897877,
+  "tests/e2e/test_dock_playground_handles.py::test_stack_minimize_all_and_expand_all[chromium]": 0.9945957090239972,
+  "tests/e2e/test_dock_playground_handles.py::test_stack_minimize_restores_previous_mix[chromium]": 1.1499660420231521,
+  "tests/e2e/test_dock_playground_handles.py::test_top_edge_resize_keeps_bottom_fixed[chromium]": 0.7188291661441326,
+  "tests/e2e/test_dock_playground_handles.py::test_top_left_corner_resize[chromium]": 0.6987448344007134,
   "tests/e2e/test_dock_playground_panels.py::test_dock_beside_fullbleed_main_panel_does_not_merge_into_area[chromium]": 1.18175800004974,
   "tests/e2e/test_dock_playground_panels.py::test_floating_stack_divider_resizes_both_directions[chromium]": 1.6986137079074979,
   "tests/e2e/test_dock_playground_panels.py::test_main_panel_click_header_toggles_collapsed[chromium]": 0.5456970401573926,
@@ -122,10 +123,6 @@
   "tests/e2e/test_dock_playground_quirks.py::test_plus_drag_tears_out_expanded[chromium]": 1.8408049591816962,
   "tests/e2e/test_dock_playground_quirks.py::test_region_resize_ignores_minimized_columns[chromium]": 2.80284612509422,
   "tests/e2e/test_dock_playground_quirks.py::test_tab_title_and_keyboard_activation[chromium]": 0.822091166395694,
-  "tests/e2e/test_dock_playground_stack_and_top.py::test_stack_minimize_all_and_expand_all[chromium]": 0.9945957090239972,
-  "tests/e2e/test_dock_playground_stack_and_top.py::test_stack_minimize_restores_previous_mix[chromium]": 1.1499660420231521,
-  "tests/e2e/test_dock_playground_stack_and_top.py::test_top_edge_resize_keeps_bottom_fixed[chromium]": 0.7188291661441326,
-  "tests/e2e/test_dock_playground_stack_and_top.py::test_top_left_corner_resize[chromium]": 0.6987448344007134,
   "tests/e2e/test_dock_playground_state.py::test_collapsed_vertical_stack_sibling_expands[chromium]": 1.2328576671425253,
   "tests/e2e/test_dock_playground_state.py::test_container_resize_keeps_floating_corners_onscreen[chromium]": 0.7413065831642598,
   "tests/e2e/test_dock_playground_state.py::test_floating_resize_grip_hidden_when_minimized[chromium]": 0.39001591713167727,
@@ -136,12 +133,15 @@
   "tests/e2e/test_dock_playground_state.py::test_rapid_sequential_tab_reorders_no_stuck_state[chromium]": 2.114137751283124,
   "tests/e2e/test_dock_playground_state.py::test_resize_storm_strands_nothing[chromium]": 0.49084049882367253,
   "tests/e2e/test_dock_playground_state.py::test_viewport_resize_mid_drag_no_desync[chromium]": 0.6271454168017954,
+  "tests/e2e/test_dock_playground_width.py::test_dock_above_single_panel_splits_height_evenly[chromium]": 1.1966556245461106,
+  "tests/e2e/test_dock_playground_width.py::test_dock_above_two_columns_spans_and_preserves_widths[chromium]": 1.688972083851695,
   "tests/e2e/test_dock_playground_width.py::test_dock_new_column_keeps_existing_and_defaults_new[chromium]": 1.1409784168936312,
   "tests/e2e/test_dock_playground_width.py::test_dock_then_undock_roundtrips_docked_width[chromium]": 2.029562624869868,
   "tests/e2e/test_dock_playground_width.py::test_no_width_jump_at_drag_start[chromium]": 1.1170597500167787,
   "tests/e2e/test_dock_playground_width.py::test_removal_float_out_preserves_sibling_width[chromium]": 1.1306902510114014,
   "tests/e2e/test_dock_playground_width.py::test_reserved_divider_resizes_with_minimized_column[chromium]": 1.539775124983862,
   "tests/e2e/test_dock_playground_width.py::test_sole_dock_adopts_region_width_not_float_width[chromium]": 1.9652793747372925,
+  "tests/e2e/test_dock_playground_width.py::test_vertical_cascade_pushes_through[chromium]": 1.119480666005984,
   "tests/e2e/test_floating_panel.py::test_drag_moves_floating_panel[chromium]": 1.5811286251991987,
   "tests/e2e/test_floating_panel.py::test_drag_to_left_edge_docks[chromium]": 1.5851812500040978,
   "tests/e2e/test_floating_panel.py::test_drag_to_right_edge_docks[chromium]": 1.6009116237983108,
